[refactor #83] 유저에게 입력받는 기능 내 필수값 유효성 검증 추가
