--- src/main/java/com/wellcome/WellcomeBE/domain/tripPlanPlace/dto/request/TripPlanPlaceRequest.java
@@ -1,8 +1,0 @@
-package com.wellcome.WellcomeBE.domain.tripPlanPlace.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class TripPlanPlaceRequest {
-    Long wellnessInfoId;
-}
